feat: Add basic support for Docker
This was added in an attempt to debug why `cargo test` succeeds on my machine and fails on GitHub's workflow runner. It didn't help. The testd on GitHub still fail.

Now I'm thinking that the problem is related to my video/ folder being in Git LFS. However, I don't want to lose this Dockerfile, so I'm going to commit it as a useful artifact to maybe help some future debugging session.